// user_service/src/generated/user/v1/user.ts
export enum Gender {
  GENDER_UNSPECIFIED = 0,
  GENDER_MALE = 1,
  GENDER_FEMALE = 2,
  UNRECOGNIZED = -1,
}

export enum RelocationReadiness {
  RELOCATION_READINESS_UNSPECIFIED = 0,
  RELOCATION_READINESS_NOT_READY = 1,
  RELOCATION_READINESS_POSSIBLE = 2,
  RELOCATION_READINESS_READY = 3,
  UNRECOGNIZED = -1,
}

export enum ProficiencyLevel {
  PROFICIENCY_LEVEL_UNSPECIFIED = 0,
  PROFICIENCY_LEVEL_BEGINNER_A1 = 1,
  PROFICIENCY_LEVEL_ELEMENTARY_A2 = 2,
  PROFICIENCY_LEVEL_INTERMEDIATE_B1 = 3,
  PROFICIENCY_LEVEL_UPPER_INTERMEDIATE_B2 = 4,
  PROFICIENCY_LEVEL_ADVANCED_C1 = 5,
  PROFICIENCY_LEVEL_NATIVE_C2 = 6,
  UNRECOGNIZED = -1,
}

export enum EducationLevel {
  EDUCATION_LEVEL_UNSPECIFIED = 0,
  EDUCATION_LEVEL_SECONDARY = 1,
  EDUCATION_LEVEL_SPECIAL_SECONDARY = 2,
  EDUCATION_LEVEL_UNFINISHED_HIGHER = 3,
  EDUCATION_LEVEL_BACHELOR = 4,
  EDUCATION_LEVEL_MASTER = 5,
  EDUCATION_LEVEL_HIGHER = 6,
  EDUCATION_LEVEL_PHD = 7,
  UNRECOGNIZED = -1,
}

export enum ResumeStatus {
  RESUME_STATUS_UNSPECIFIED = 0,
  RESUME_STATUS_DRAFT = 1,
  RESUME_STATUS_PUBLISHED = 2,
  RESUME_STATUS_ARCHIVED = 3,
  UNRECOGNIZED = -1,
}

export enum ResumeVisibility {
  RESUME_VISIBILITY_UNSPECIFIED = 0,
  RESUME_VISIBILITY_PUBLIC = 1,
  RESUME_VISIBILITY_REGISTERED_EMPLOYERS = 2,
  RESUME_VISIBILITY_PRIVATE = 3,
  UNRECOGNIZED = -1,
}

export enum SearchStatus {
  SEARCH_STATUS_UNSPECIFIED = 0,
  SEARCH_STATUS_ACTIVE_SEARCH = 1,
  SEARCH_STATUS_OPEN_TO_OFFERS = 2,
  SEARCH_STATUS_NOT_SEARCHING = 3,
  SEARCH_STATUS_FOUND_JOB = 4,
  UNRECOGNIZED = -1,
}

export enum BusinessTripReadiness {
  BUSINESS_TRIP_READINESS_UNSPECIFIED = 0,
  BUSINESS_TRIP_READINESS_NEVER = 1,
  BUSINESS_TRIP_READINESS_SOMETIMES = 2,
  BUSINESS_TRIP_READINESS_READY = 3,
  UNRECOGNIZED = -1,
}

export enum EmploymentType {
  EMPLOYMENT_TYPE_UNSPECIFIED = 0,
  EMPLOYMENT_TYPE_FULL_TIME = 1,
  EMPLOYMENT_TYPE_PART_TIME = 2,
  EMPLOYMENT_TYPE_PROJECT = 3,
  EMPLOYMENT_TYPE_INTERNSHIP = 4,
  EMPLOYMENT_TYPE_TEMPORARY = 5,
  EMPLOYMENT_TYPE_VOLUNTEER = 6,
  UNRECOGNIZED = -1,
}

export enum WorkSchedule {
  WORK_SCHEDULE_UNSPECIFIED = 0,
  WORK_SCHEDULE_FULL_DAY = 1,
  WORK_SCHEDULE_SHIFT = 2,
  WORK_SCHEDULE_FLEXIBLE = 3,
  WORK_SCHEDULE_ROTATION = 4,
  UNRECOGNIZED = -1,
}

export enum WorkFormat {
  WORK_FORMAT_UNSPECIFIED = 0,
  WORK_FORMAT_ONSITE = 1,
  WORK_FORMAT_REMOTE = 2,
  WORK_FORMAT_HYBRID = 3,
  WORK_FORMAT_FIELD = 4,
  UNRECOGNIZED = -1,
}